Attempt at splitting up the code
